Refactor ontology viewer page
