+++ src/components/viewers/Page.js
@@ -1,0 +1,58 @@
+import React, { useEffect, useState } from 'react';
+import Layout from '@theme/Layout';
+import {
+  BrowserRouter as Router,
+  useLocation
+} from "react-router-dom";
+import ResourceView from './ResourceView';
+import ClassList from './ClassList';
+import ScrollToTop from '../ScrollToTop';
+
+function useQuery() {
+    return new URLSearchParams(useLocation().search);
+}
+
+function PageComponent({settings}) {
+    let query = useQuery();
+    const uri = query.get('uri')
+
+    // Reason we serialise fetchOptions is to avoid infinite loop.
+    // See https://github.com/vercel/swr/issues/345
+    // const { data, error } = useSWR([endpoint, JSON.stringify(fetchOptions)], fetcher)
+
+    const [selectedClass, setSelectedClass] = useState(null)
+
+    useEffect(() => {
+        if (uri) {
+            setSelectedClass(uri)
+        }
+    }, [uri])
+
+    return (
+        <main className="margin-vert--lg">
+            <div className="row">
+                <div className="col col--2">
+                    <ClassList settings={settings} />
+                </div>
+
+                <div className="col col--10">
+                    <ResourceView resourceUri={selectedClass} settings={settings} />
+                </div>
+            </div>
+        </main>
+    )
+}
+
+export default function Page({settings}) {
+    if (typeof window === "undefined") {
+        return <></>
+    }
+
+    return <Layout title="TERN Ontology viewer">
+        <Router>
+            <ScrollToTop>
+                <PageComponent settings={settings} />
+            </ScrollToTop>
+        </Router>
+    </Layout>
+}
