+++ src/pages/viewers/tern-ontology/_settings.js
@@ -1,6 +1,7 @@
 import * as queries from './_queries'
 
 const settings = {
+    title: 'TERN Ontology viewer',
     pageRoute: '/viewers/tern-ontology',
     endpoint: 'https://graphdb.tern.org.au/repositories/knowledge_graph_core',
     queries: queries
